Add support for automatic route insertion/removal
This saves the operator/system having to implement this externally with
shell scripts.

Note: The shutdown ordering is a bit hacky. I had to replace the use of
log.Fatal() with log.Print() because for some reason shutting down the
listener(s) is unclean which prevents the removal of the route. Needs
investigation. The listeners should shutdown without error.

Signed-off-by: Joe Groocock <me@frebib.net>